+++ CHANGELOG.md
@@ -4,9 +4,18 @@ All notable changes to this project will be documented in this file.
 
 ## [0.2.1]
 
+### Fixed
+
+* Using `from __future__ import annotations` caused issues with dataclass field type resolution. This has been fixed to ensure compatibility.
+* Generation of yaml for dictionary types with arbitrary keys and typed values now working.
+
 ### Changed
 
 * Moved `Primitives` and `primitive_types` to `eyconf.constants` for better modularity.
+
+### Added
+
+* More examples in the quickstart guide.
 
 
 ## [0.2.0] - 2025-10-23